Updated to Forge 859
ran Source > Clean Up on most of the files

+++ bedcraftbeyond/client/render/BlockBedRenderer.java
@@ -1,20 +1,14 @@
 package zornco.bedcraftbeyond.client.render;
 
 import net.minecraft.block.Block;
-import net.minecraft.block.BlockHalfSlab;
-import net.minecraft.block.BlockStairs;
-import net.minecraft.block.material.Material;
 import net.minecraft.client.renderer.RenderBlocks;
 import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.util.Direction;
 import net.minecraft.util.Icon;
 import net.minecraft.world.IBlockAccess;
 
-import org.lwjgl.opengl.GL11;
-
 import zornco.bedcraftbeyond.BedCraftBeyond;
 import zornco.bedcraftbeyond.blocks.BlockColoredBed;
-import zornco.bedcraftbeyond.blocks.BlockRug;
 import cpw.mods.fml.client.registry.ISimpleBlockRenderingHandler;
 
 public class BlockBedRenderer implements ISimpleBlockRenderingHandler {
@@ -43,15 +37,15 @@ public class BlockBedRenderer implements ISimpleBlockRenderingHandler {
 		tessellator.setColorOpaque_F(f, f, f);
 		Icon icon = block.getIcon(0, world.getBlockMetadata(x, y, z), 0, world, x, y, z);
 		if (renderer.hasOverrideBlockTexture()) icon = renderer.overrideBlockTexture; //BugFix Proper breaking texture on underside
-		double d0 = (double)icon.getMinU();
-		double d1 = (double)icon.getMaxU();
-		double d2 = (double)icon.getMinV();
-		double d3 = (double)icon.getMaxV();
-		double d4 = (double)x + renderer.renderMinX;
-		double d5 = (double)x + renderer.renderMaxX;
-		double d6 = (double)y + renderer.renderMinY + 0.1875D;
-		double d7 = (double)z + renderer.renderMinZ;
-		double d8 = (double)z + renderer.renderMaxZ;
+		double d0 = icon.getMinU();
+		double d1 = icon.getMaxU();
+		double d2 = icon.getMinV();
+		double d3 = icon.getMaxV();
+		double d4 = x + renderer.renderMinX;
+		double d5 = x + renderer.renderMaxX;
+		double d6 = y + renderer.renderMinY + 0.1875D;
+		double d7 = z + renderer.renderMinZ;
+		double d8 = z + renderer.renderMaxZ;
 		tessellator.addVertexWithUV(d4, d6, d8, d0, d3);
 		tessellator.addVertexWithUV(d4, d6, d7, d0, d2);
 		tessellator.addVertexWithUV(d5, d6, d7, d1, d2);
@@ -64,10 +58,10 @@ public class BlockBedRenderer implements ISimpleBlockRenderingHandler {
 			tessellator.setColorOpaque_I(BlockColoredBed.getColorFromTilePerPass(world, x, y, z, i));
 			icon = block.getIcon(1, world.getBlockMetadata(x, y, z), i, world, x, y, z);
 			if (renderer.hasOverrideBlockTexture()) icon = renderer.overrideBlockTexture; //BugFix Proper breaking texture on underside
-			d0 = (double)icon.getMinU();
-			d1 = (double)icon.getMaxU();
-			d2 = (double)icon.getMinV();
-			d3 = (double)icon.getMaxV();
+			d0 = icon.getMinU();
+			d1 = icon.getMaxU();
+			d2 = icon.getMinV();
+			d3 = icon.getMaxV();
 			d4 = d0;
 			d5 = d1;
 			d6 = d2;
@@ -103,11 +97,11 @@ public class BlockBedRenderer implements ISimpleBlockRenderingHandler {
 				d11 = d2;
 			}
 
-			double d12 = (double)x + renderer.renderMinX;
-			double d13 = (double)x + renderer.renderMaxX;
-			double d14 = (double)y + renderer.renderMaxY;
-			double d15 = (double)z + renderer.renderMinZ;
-			double d16 = (double)z + renderer.renderMaxZ;
+			double d12 = x + renderer.renderMinX;
+			double d13 = x + renderer.renderMaxX;
+			double d14 = y + renderer.renderMaxY;
+			double d15 = z + renderer.renderMinZ;
+			double d16 = z + renderer.renderMaxZ;
 			tessellator.addVertexWithUV(d13, d14, d16, d8, d10);
 			tessellator.addVertexWithUV(d13, d14, d15, d4, d6);
 			tessellator.addVertexWithUV(d12, d14, d15, d5, d7);
@@ -144,7 +138,7 @@ public class BlockBedRenderer implements ISimpleBlockRenderingHandler {
 			for (int i = 0; i < 3; i++) {
 				tessellator.setColorOpaque_I(BlockColoredBed.getColorFromTilePerPass(world, x, y, z, i));
 				renderer.flipTexture = b0 == 2;
-				renderer.renderFaceZNeg(block, (double)x, (double)y, (double)z, block.getIcon(2, world.getBlockMetadata(x, y, z), i, world, x, y, z));
+				renderer.renderFaceZNeg(block, x, y, z, block.getIcon(2, world.getBlockMetadata(x, y, z), i, world, x, y, z));
 			}
 		}
 		if (k1 != 3 && (renderer.renderAllFaces || block.shouldSideBeRendered(renderer.blockAccess, x, y, z + 1, 3)))
@@ -154,7 +148,7 @@ public class BlockBedRenderer implements ISimpleBlockRenderingHandler {
 			for (int i = 0; i < 3; i++) {
 				tessellator.setColorOpaque_I(BlockColoredBed.getColorFromTilePerPass(world, x, y, z, i));
 				renderer.flipTexture = b0 == 3;
-				renderer.renderFaceZPos(block, (double)x, (double)y, (double)z, block.getIcon(3, world.getBlockMetadata(x, y, z), i, world, x, y, z));
+				renderer.renderFaceZPos(block, x, y, z, block.getIcon(3, world.getBlockMetadata(x, y, z), i, world, x, y, z));
 			}
 		}
 
@@ -165,7 +159,7 @@ public class BlockBedRenderer implements ISimpleBlockRenderingHandler {
 			for (int i = 0; i < 3; i++) {
 				tessellator.setColorOpaque_I(BlockColoredBed.getColorFromTilePerPass(world, x, y, z, i));
 				renderer.flipTexture = b0 == 4;
-				renderer.renderFaceXNeg(block, (double)x, (double)y, (double)z, block.getIcon(4, world.getBlockMetadata(x, y, z), i, world, x, y, z));
+				renderer.renderFaceXNeg(block, x, y, z, block.getIcon(4, world.getBlockMetadata(x, y, z), i, world, x, y, z));
 			}
 		}
 
@@ -176,7 +170,7 @@ public class BlockBedRenderer implements ISimpleBlockRenderingHandler {
 			for (int i = 0; i < 3; i++) {
 				tessellator.setColorOpaque_I(BlockColoredBed.getColorFromTilePerPass(world, x, y, z, i));
 				renderer.flipTexture = b0 == 5;
-				renderer.renderFaceXPos(block, (double)x, (double)y, (double)z, block.getIcon(5, world.getBlockMetadata(x, y, z), i, world, x, y, z));
+				renderer.renderFaceXPos(block, x, y, z, block.getIcon(5, world.getBlockMetadata(x, y, z), i, world, x, y, z));
 			}
 		}
